Wrap another line in `unsafe`
It's generally better to include everything that makes a particular `unsafe` block safe within the block. Just wrapping a single function call in `unsafe` doesn't give you information about what you've done to make it safe.

In this case, we ensure that the string isn't null, and so that makes it safe to call `from_ptr`.